requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==24.1.0
    # via aiogram
aiogram==3.17.0
    # via -r requirements.in
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via aiogram
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==25.1.0
    # via aiohttp
certifi==2025.1.31
    # via aiogram
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.10.6
    # via aiogram
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   aiogram
    #   pydantic
    #   pydantic-core
yarl==1.18.3
    # via aiohttp
